pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "AI job-search assistant built with LangChain, LangSmith, FastAPI, and Postgres."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic",
    "deepagents",
    "fastapi",
    "langchain",
    "langchain-openai",
    "langchain-tavily>=0.2.18",
    "langgraph-cli[inmem]>=0.4.24",
    "langsmith",
    "psycopg[binary]",
    "pydantic-settings",
    "sqlmodel",
    "uvicorn[standard]",
]

[project.scripts]
api = "job_search_agent.dev:run_api"

[project.optional-dependencies]
dev = [
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_search_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

